requirements.txt
cloudsigma==0.3
hurry.filesize==0.9
prettytable==0.7.2
